chg: dev: refactor ``DatePicker`` related intelligence out to dedicated service
